% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_EPA.R
\name{align_and_trim_query}
\alias{align_and_trim_query}
\title{Align and trim query sequence}
\usage{
align_and_trim_query(
  seqs,
  tree,
  ref,
  model,
  save.paths,
  numCores = 1,
  max.out = 1
)
}
\description{
Align and trim query sequence
}
