% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{check_partition}
\alias{check_partition}
\title{Parse grouping information}
\usage{
check_partition(tree, dAIC, group)
}
\description{
This function parse grouping information of nodes to contiguous neighborhoods
}
