% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{get_connected_nodes}
\alias{get_connected_nodes}
\title{Grouping connected nodes into neighborhoods}
\usage{
get_connected_nodes(tree, node, group)
}
\description{
This function finds the maximal neighborhood that includes the focal node
}
