% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{get_partition_AIC}
\alias{get_partition_AIC}
\title{Calculate AIC after partition}
\usage{
get_partition_AIC(tree, trait, models, group)
}
\description{
This function calculates the AIC of the heterogeneous model
}
