% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{get_partition_transition}
\alias{get_partition_transition}
\title{Get number of transitions}
\usage{
get_partition_transition(tree, group)
}
\description{
This function calculates the number of transitions between groups
}
