% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{partition_phylogeny_by_AIC_single}
\alias{partition_phylogeny_by_AIC_single}
\title{Binary partition of a phylogeny}
\usage{
partition_phylogeny_by_AIC_single(tree, dAIC, flip = FALSE)
}
\description{
This function partition a phylogeny given the delta AIC in each node
}
