% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_wrapper.R
\name{predict_16SGCN_from_jplace}
\alias{predict_16SGCN_from_jplace}
\title{Predict 16S rRNA GCN from sequence placements}
\usage{
predict_16SGCN_from_jplace(jplace, numCores = 1, save2file = FALSE)
}
\description{
A wrapper function to the prediction from placements
}
