% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_wrapper.R
\name{predict_16SGCN_from_sequences}
\alias{predict_16SGCN_from_sequences}
\title{Predict 16S rRNA GCN from sequences}
\usage{
predict_16SGCN_from_sequences(seqs)
}
\description{
A wrapper function to "one-click-and-run" the default pipeline
}
