% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_data.R
\docType{data}
\name{RasperGade16S.GG.13.8.mask.keys}
\alias{RasperGade16S.GG.13.8.mask.keys}
\title{GreenGene 13.8 16S rRNA gene mask for RasperGade16S}
\format{
A character vector of 1287 elements; each element has 4938 characters:
\describe{
\item{values}{16S rRNA gene sequence at each position}
\item{seq_id}{the attribute that marks the identity of each 16S rRNA gene sequence}
}
}
\source{
\url{https://github.com/wu-lab-uva/16S-rRNA-GCN-Predcition}
}
\usage{
RasperGade16S.GG.13.8.mask.keys
}
\description{
A character vector of 1287 elements
}
\keyword{datasets}
