% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_data.R
\docType{data}
\name{RasperGade16S.refGCN}
\alias{RasperGade16S.refGCN}
\title{Default reference GCN for RasperGade16S}
\format{
A named vector with 6408 elements:
\describe{
\item{values}{16S rRNA GCN of reference genomes}
\item{names}{reference genome names, in NCBI RefSeq accession number}
}
}
\source{
\url{https://github.com/wu-lab-uva/16S-rRNA-GCN-Predcition}
}
\usage{
RasperGade16S.refGCN
}
\description{
A named vector of 6408 elements
}
\keyword{datasets}
