% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_data.R
\docType{data}
\name{RasperGade16S.reftree}
\alias{RasperGade16S.reftree}
\title{Default reference tree for RasperGade16S}
\format{
A phylo-class object with 6408 tips, binary and rooted:
\describe{
\item{tip.label}{reference genome names, in NCBI RefSeq accession number}
\item{edge.length}{phylogenetic distance between nodes, in substitutions/site}
\item{edge}{parent node and descendant node of an edge}
\item{Nnode}{the number of nodes}
}
}
\source{
\url{https://github.com/wu-lab-uva/16S-rRNA-GCN-Predcition}
}
\usage{
RasperGade16S.reftree
}
\description{
A phylo-class object containing 6408 reference tips/genomes
}
\keyword{datasets}
