% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_data.R
\docType{data}
\name{RasperGade16S.SILVA.GCN}
\alias{RasperGade16S.SILVA.GCN}
\title{Predicted SILVA 132 GCN (Bacteria only) by RasperGade16S}
\format{
A named numeric vector with 592605 elements:
\describe{
\item{values}{16S rRNA GCN of SILVA bacterial OTUs}
\item{names}{reference genome names, in NCBI RefSeq accession number}
}
}
\source{
\url{https://github.com/wu-lab-uva/16S-rRNA-GCN-Predcition}
}
\usage{
RasperGade16S.SILVA.GCN
}
\description{
A named vector of 592605 elements
}
\keyword{datasets}
