% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{refine_partition}
\alias{refine_partition}
\title{Refine partition by reverting groups}
\usage{
refine_partition(tree, dAIC, group)
}
\description{
This function refines neighborhoods that have worse AIC
}
