% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_partition.R
\name{rescale_tree_by_partition_and_model}
\alias{rescale_tree_by_partition_and_model}
\title{Rescale tree given a partition and models}
\usage{
rescale_tree_by_partition_and_model(
  tree,
  group,
  models,
  epsilon.as.PE = FALSE,
  min.value = 1e-06
)
}
\description{
This function rescales a tree so that a homogeneous model can apply
}
