% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_EPA.R
\name{trim_sequence_with_mask}
\alias{trim_sequence_with_mask}
\title{Trim 16S sequence based on GreenGene 13.8 16S mask}
\usage{
trim_sequence_with_mask(align, trimmed.align)
}
\description{
Trim 16S sequence based on GreenGene 13.8 16S mask
}
