% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RasperGade16S_EPA.R
\name{insert_query_with_EPA}
\alias{insert_query_with_EPA}
\title{Run EPA from R terminal}
\usage{
insert_query_with_EPA(
  seqs,
  tree,
  ref.seqs,
  model,
  save.path,
  numCores = 0,
  max.out = 99,
  intern = TRUE
)
}
\description{
Calls EPA-ng and insert query sequences into the reference
}
